+++ app/(app)/recipes/recipe/[id].tsx
@@ -1,7 +1,7 @@
-import { useEffect, useState } from 'react';
-import { Text, ScrollView } from 'react-native';
+import { ReactNode, useEffect, useState } from 'react';
+import { Text, ScrollView, View } from 'react-native';
 import { useLocalSearchParams } from 'expo-router';
-import { commonStyles } from '@/style/commonStyles';
+import { colors, commonStyles } from '@/style/commonStyles';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { textStyles } from '@/style/textStyles';
 import { useRecipes } from '@/hooks/useRecipes';
@@ -26,9 +26,37 @@ export default function recipe() {
   }, [id]);
 
   const MarkdownText = ({ content }: { content: string }) => {
+    const customRenderRules = {
+      bullet_list: (node: any, children: ReactNode[]) => (
+        <View style={{ paddingLeft: 20 }}>
+          {children}
+        </View>
+      ),
+      ordered_list: (node: any, children: ReactNode[]) => (
+        <View style={{ paddingLeft: 20 }}>
+          {children}
+        </View>
+      ),
+      list_item: (node: any, children: ReactNode[]) => (
+        <Text style={{ color: colors.sg2 }}>
+          {node.type === 'bullet_list' ? '• ' : `${node.index + 1}. `}
+          {children}
+        </Text>
+      ),
+    };
+
     return (
       <Text style={textStyles.textLg}>
-        <Markdown>{content}</Markdown>
+        <Markdown
+          style={{
+            text: { color: colors.sg2 },
+            paragraph: { color: colors.sg2 },
+            heading: { color: colors.sg2 },
+          }}
+          rules={customRenderRules}
+        >
+          {content}
+        </Markdown>
       </Text>
     );
   }
@@ -56,9 +84,9 @@ export default function recipe() {
                   }}
                 />
 
+                {/* <Text style={textStyles.textLg}>{recipe.description}</Text> */}
                 {/* @ts-ignore */}
-                <Text style={textStyles.textLg}>{recipe.description}</Text>
-                {/* <MarkdownText content={recipe.description} /> */}
+                <MarkdownText content={recipe.description} />
 
               </ScrollView>
             ) : (
@@ -72,3 +100,17 @@ export default function recipe() {
     </SafeAreaProvider>
   );
 }
+
+// const styles = StyleSheet.create({
+//   container: {
+//     marginTop: 50,
+//   },
+//   bigBlue: {
+//     color: 'blue',
+//     fontWeight: 'bold',
+//     fontSize: 30,
+//   },
+//   red: {
+//     color: 'red',
+//   },
+// });